+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Lower versions of Gradle might be supported but are not tested.
     In particular this the `generateHelpMojo` property was removed.
     Users should configure `helpMojoPackage` with the desired target package for the generated help mojo.
 * [Fixed] [8](https://github.com/britter/maven-plugin-development/issues/278) Support configuration cache.
+* [Fixed] [38](https://github.com/britter/maven-plugin-development/issues/278) Don't use Property for lazy types.
+    Instead of declaring fields of type `Property<FileCollection>` the `GenerateMavenPluginDescriptorTask` not declares fields of type `ConfigurableFileCollection`, which is more idiomatic.
 
 ## Version 0.4.3